Tests running with patched pip-install
