+++ .github/workflows/main.yml
@@ -6,6 +6,7 @@ on:
   push:
     branches:
       - main
+      - alexej/ENG-234-zenml-integration-install
 
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
